Προετοιμασία για τα κριτήρια εμφάνισης
